researcher-smolagent/requirements.txt
# Core framework
smolagents==1.8.0
chainlit==1.3.1
openai==1.61.1
python-dotenv==1.0.0

# Web scraping and search
duckduckgo-search==6.3.5
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
trafilatura==1.12.2

# Document processing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2

# Data handling
pandas==2.2.3

# Async support
aiohttp==3.10.11
